refactor(cli): extract manifest MCP domain
